requirements.txt
asgiref==3.4.1
cffi==1.15.0
chardet==4.0.0
click==8.0.3
cryptography==36.0.0
Django==3.2.9
joblib==1.1.0
nltk==3.6.5
pdfminer.six==20211012
pycparser==2.21
PyPDF2==1.26.0
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
textblob==0.17.1
tqdm==4.62.3
